+++ src/subLayers/geoTileAggregation.js
@@ -63,6 +63,13 @@ class GeoTileAggregation {
     }
 
     updateQuery(query, aggField, precision=0) {
+
+        if(undefined === query.body.query.bool.filter){
+            if(this.aggConfig.options.defaultExtent){
+                query.body.query.bool.filter = [this.aggConfig.options.defaultExtent];
+            }
+        }
+
         let aggs = {
             agg: {
                 geotile_grid: {field: aggField, precision, size: this.indexConfig.maxResults},